requirements.txt
together==1.2.11
python-dotenv==1.0.1
pinecone==5.1.0
pyyaml==6.0.2
requests==2.32.3
beautifulsoup4==4.12.3 
sentence-transformers==3.0.1
langchain==0.2.16
lxml==5.3.0
wikitextparser==0.56.2
tiktoken==0.7.0
